test(wasm-utxo): remove deprecated fromHalfSignedLegacyTransaction tests
Remove tests for deprecated `fromHalfSignedLegacyTransaction` method.
Retain tests for `fromNetworkFormat` which supersedes it.

Consolidate test file and reduce redundancy by combining half-signed
and full-signed test cases into a single parameterized helper.

Issue: BTC-2650
